refactor: move FooterIconsColumn and Footer to components
