+++ packages/components/src/organisms/FFooter/FFooter.constants.ts
@@ -1,0 +1,11 @@
+export const listService: string[] = [
+  "Conta corrente",
+  "Conta PJ",
+  "Cartão de crédito",
+];
+
+export const listContact: string[] = [
+  "0800 004 250 08",
+  "meajuda@bytebank.com.br",
+  "ouvidoria@bytebank.com.br",
+];
